+++ app/pb_kits/playbook/pb_weekday_stacked/docs/_weekday_stacked_compact.jsx
@@ -1,0 +1,48 @@
+import React from 'react'
+import {
+  Flex,
+  // FlexItem,
+  WeekdayStacked,
+} from '../..'
+
+const WeekdayStackedCompact = () => (
+  <Flex spacing="evenly">
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/4')}
+        dayOnly
+    />
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/5')}
+        dayOnly
+    />
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/6')}
+        dayOnly
+    />
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/7')}
+        dayOnly
+    />
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/8')}
+        dayOnly
+    />
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/9')}
+        dayOnly
+    />
+    <WeekdayStacked
+        compact
+        date={new Date('2019/11/10')}
+        dayOnly
+    />
+  </Flex>
+)
+
+export default WeekdayStackedCompact