+++ app/pb_kits/playbook/pb_weekday_stacked/docs/_weekday_stacked_dark.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import {
+  WeekdayStacked,
+} from '../../'
+
+const WeekdayStackedDark = () => (
+  <div>
+    <WeekdayStacked
+        dark
+    />
+  </div>
+)
+
+export default WeekdayStackedDark